Switch nLen to size_t and remove a cast of the output of strlen in OGRAVCLayer::TranslateTableFields().
git-svn-id: https://svn.osgeo.org/gdal/trunk@32124 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/avc/ogravclayer.cpp
@@ -568,9 +568,8 @@ int OGRAVCLayer::TranslateTableFields( OGRFeature *poFeature,
             if (nType == AVC_FT_CHAR)
             {
                 /* Remove trailing spaces in char fields */
-                int nLen = static_cast<int>(
-                    strlen( reinterpret_cast<char *>(
-                        pasFields[iField].pszStr) ) );
+                size_t nLen = strlen( reinterpret_cast<char *>(
+                    pasFields[iField].pszStr) );
                 while (nLen > 0 && pasFields[iField].pszStr[nLen-1] == ' ')
                     nLen--;
                 pasFields[iField].pszStr[nLen] = '\0';
